Reuse product component
